fix: make conditions optional
